Extract Suggestions out of SuggestionListView
