+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/suggestion/internal/SuggestionListUi.kt
@@ -1,10 +1,10 @@
 package io.getstream.chat.android.ui.suggestion.internal
 
-import io.getstream.chat.android.ui.suggestion.list.SuggestionListView
+import io.getstream.chat.android.ui.suggestion.Suggestions
 
 internal interface SuggestionListUi {
 
-    fun showSuggestionList(suggestions: SuggestionListView.Suggestions)
+    fun showSuggestionList(suggestions: Suggestions)
 
     fun hideSuggestionList()
